Addition of 2 functional group factories to allow for Alkene Metathesis and Disulfide Exchange
Addition of 2 new Function Group Factories (FGF). these are the Alkene Metathesis Factory and the Disulfide Exchange Factory. They create .Alkene and .Disulfide instances so no new FGs needed to be added.

Works similar to a FGF except has a new variable upon initialisation which is largest_chain which is set to True by default. This will assign the bonders and deleters according to what would delete the fewest atoms. This is done by iterating over the molecule's connected atoms from the stated deleter atom using the bonds. This is done via the new utility function, _find_R_ids.

If a user needs to explicitly state the bonders, they need to set largest_chain to false, and then state the bonders in the usual format, except each alkene will correspond to a single integer in bonders. Therefore, a molecule with 5 alkenes in it would require bonders to have 5 integers in a tuple, ordered according to the order the alkenes are assigned (based on atom_ids). These integers will be either 0 or 3, corresponding to the functional group string in the docs. Examples are provided in the doc string.